test(activities): separate creation and listing e2e tests suits
